+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.1] - 2023-02-08
+### Fixed
+- When decoding enums, accept both enum member and screaming-snake case.
+  This makes it easier to transition between them.
+  When `-version=strict` is set, only the default format is accepted.
+
 ## [1.9.0] - 2023-01-11
 ### Added
 - Interpret boilerplate's `@AliasThis` equivalent to `alias this` in JSON encoding/decoding.